requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-environ==0.4.5
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
environ==1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==21.0
pathspec==0.9.0
platformdirs==2.2.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
tomli==1.2.1
typing-extensions==3.10.0.1
uritemplate==3.0.1
urllib3==1.26.6
